clean-up location prices code - remove duplication
