+++ src/location_prices/LocationPricesList.tsx
@@ -6,7 +6,14 @@ import {
     useRecordContext,
     ListProps,
     BooleanField,
+    TopToolbar,
+    ExportButton,
+    Button,
 } from 'react-admin';
+
+import { Link } from 'react-router-dom';
+import AddCircle from '@mui/icons-material/AddCircleOutline';
+import { Stack } from '@mui/material';
 
 import { CurrencyField } from '../CurrencyField';
 
@@ -63,5 +70,78 @@ export const LocationPricesList = (props: Omit<ListProps, 'children'>) => {
                 <ValueField source="market_price_fix" label="Fixed Market" />
             </Datagrid>
         </List>
+    );
+};
+
+export const LocationTabPriceList = (props: {
+    pathname: string;
+    company_id: string | number;
+    location_id?: string | number;
+}) => {
+    const { pathname, company_id, location_id } = props;
+
+    const filter = { company_id, location_id };
+    if (filter.location_id === undefined) delete filter.location_id;
+
+    return (
+        <LocationPricesList
+            resource="location_prices"
+            disableSyncWithLocation={true}
+            filter={filter}
+            actions={
+                <TopToolbar>
+                    <CreateRelatedLocation
+                        pathname={pathname}
+                        company_id={company_id}
+                        location_id={location_id}
+                    />
+                    <ExportButton />
+                </TopToolbar>
+            }
+            empty={
+                <>
+                    <Stack
+                        direction="row"
+                        justifyContent="flex-end"
+                        spacing={2}
+                        mt={1}
+                    >
+                        <CreateRelatedLocation
+                            pathname={pathname}
+                            company_id={company_id}
+                            location_id={location_id}
+                        />
+                    </Stack>
+                    <div>No records</div>
+                </>
+            }
+        />
+    );
+};
+
+const CreateRelatedLocation = (props: {
+    pathname: string;
+    company_id: string | number;
+    location_id?: string | number;
+}) => {
+    const { pathname, company_id, location_id } = props;
+
+    const state = pathname
+        ? {
+              from: pathname,
+              redirect_on_save: pathname,
+              record: { company_id: company_id, location_id },
+          }
+        : undefined;
+
+    return (
+        <Button
+            component={Link}
+            to="/location_prices/create"
+            state={state}
+            label="Add Sales Price"
+        >
+            <AddCircle />
+        </Button>
     );
 };
