+++ src/location_prices/index.ts
@@ -1,6 +1,8 @@
 import { LocationPricesList } from './LocationPricesList';
 import { LocationPricesCreate } from './LocationPricesCreate';
 
+export { LocationTabPriceList } from './LocationPricesList';
+
 export default {
     list: LocationPricesList,
     create: LocationPricesCreate,
